requirements.txt
numpy
playsound
pydub
pyworld
requests
scipy
sounddevice
SoundFile
starlette
torch
torchaudio
tqdm
scikit-maad
praat-parselmouth
librosa
torchvision
fairseq